requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-KVSession-fork==0.6.4
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
geoip2==2.9.0
python-dotenv==1.0.1
gunicorn==23.0.0
Flask-Limiter==3.9.2